Projects/port_tester/requirements.txt
# Required packages
google-generativeai>=0.3.0

# Standard library modules used:
# socket, datetime, concurrent.futures, json, os, pathlib
